+++ js/comment.js
@@ -1,0 +1,41 @@
+import {getrandomNumber} from '../js/util.js';
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAME = [
+  'Виктор',
+  'Мария',
+  'Иван',
+  'Наталья',
+  'Станислав',
+  'Маргарита',
+  'Виктор',
+  'Яна',
+  'Денис',
+  'Дарья',
+  'Макар',
+  'Анастасия',
+  'Егор',
+  'Татьяна'
+];
+
+const MIN_AVATAR = 1;
+
+const MAX_AVATAR = 6;
+
+const createComment = (id) => ({
+  id: id,
+  avatar: `img/avatar-${getrandomNumber (MIN_AVATAR, MAX_AVATAR)}.svg`,
+  message: MESSAGE[getrandomNumber (0, MESSAGE.length - 1)],
+  name: NAME[getrandomNumber (0, NAME.length - 1)]
+});
+
+export {createComment};
