+++ js/photos.js
@@ -1,0 +1,24 @@
+import {createComment} from '../js/comment.js';
+
+import {createPost} from '../js/post.js';
+
+const createPhotos = () => {
+  const photos = [];
+
+  for (let i = 0; i < 25; i++) {
+    const post = createPost(i + 1);
+    const comments = [];
+
+    for (let j = 0; j < 3; j++) {
+      const comment = createComment(j + 1);
+      comments.push(comment);
+    }
+
+    post.comments = comments;
+    photos.push(post);
+  }
+
+  return photos;
+};
+
+export {createPhotos};
